refactor query keys and update certificaciones components for improved structure
