--- src/modules/expedientes/editar-certificaciones/editar-certificaciones-keys.ts
@@ -1,4 +1,0 @@
-export const certificacionesQueryKeys = {
-  all: ['certificaciones'],
-  certificacionCollection: (page?: number, pageSize?: number) => ['certificaciones', page, pageSize],
-};
