--- src/modules/expedientes/expedientes-keys.ts
@@ -1,4 +1,0 @@
-export const expedientesQueryKeys = {
-  all: ['expedientes'],
-  expedienteCollection: (page?: number, pageSize?: number) => ['expedientes', page, pageSize],
-};
